Added multi-currency twig template
[WIP] Added twig template for the main menu wrap and dummy logic for the multi currency section

Added an autoloader and removed the existing autoloader form the compatibility class

Added the sub-templates for the multi-currency screen

More changes for sub-templates

Fixes for multi-currency template

Fixes for the multi-currency template

Fixes for the multi-currency template

Typo

Fixes for the multi-currency template

Multiple fixes for the multi currency screen (incl. wcml-960)

Multiple fixes for the multi currency screen (incl. wcml-960)

More fixes for the multi currency screen (incl. wcml-960)