Remove the use of `\WooCommerce_WPML::instance`
To handle this (that is, to make tests pass), I needed to fix the depndencies in `WCML_Url_Translation`.

The class was already expecting two of the needed dependencies, but they weren't used, for some reason.

I've added also the instance of `wpdb`, so to get rid of the global.

Other changes:

- In the main plugin file, removed some duplication and the use of the singleton
- In `WCML_TP_Support`, get the instance of `WooCommerce_WPML` from the global variable, and create the instance with `WCML_Factory` if the former is not set

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-1966