Do not try to translate categories when the shortcode attribute is empty
The "category" attribute is set to an empty string by default, so it
should be always set.

In Elementor's "Products" block, the filter
"woocommerce_shortcode_products_query" is applied on query args that
already contain a custom tax query for product categories, but the
relative "category" attribute is missing (we have the default empty string
instead).

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-2634