From `https://git.onthegosystems.com/wpml-shared/makefile-git-hooks` (6f95635f)
This new version of the script will properly update the version number in the main plugin file.

It relies on the JSON array stored in a `OTGS_CI_REPLACEMENTS` environment variable.

See how the variable must be set by running `deploy-update-version.js` without any argument.

For debugging purposes, the variable will be replaced with an hard-coded JSON array when using the `--debug` argument.

https://onthegosystems.myjetbrains.com/youtrack/issue/wpmlga-523