Fix issue preventing to reload `woocommerce` MO files on switch
The code removed was just causing troubles because it was unloading the
domain `woocommerce` (=> `$GLOBALS['l10n_unloaded']`). Then, after
loading it again, it should be removed from the unloaded global.
However, when the switched language did not have a valid native language file, the loading was failing and the domain was not removed from the unloaded global.

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-2918