Fix wrong email language when changing status from "On Hold" to "Processing"
Some code was removed in
https://git.onthegosystems.com/wpml/woocommerce-multilingual/commit/9a49ecb49d4b9c9879017c7f9654be6ddf6b72b7
and it seems that some other code was also interfering.

Basically, WCML should not try to unload/switch language/load WC text domain because
WPML ST is supposed to do this for any plugin. WCML is only responsible
for switching the language which is done in
`WCML_Emails::change_email_language`.

I also removed the code that was unloading/switching language/loading WC text domain in
other places.

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-2934