Prevent flushing rewrite rules in a non-default language when it's queued
When saving on WC > Settings > Products, WC sets a flag to flush the
rewrite rule on the next request. If the next request is in a
non-default language, it will break the rules. That's what we want to
avoid. The rules will be flushed on the first default language request.

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-2937