Fix fatal error when TM is disabled by Core
The fatal error is thrown on `init`.

The constant `WPML_TM_VERSION` is set, but TM is disabled by Core in
`\WPML_Plugins_Check::disable_outdated` (done on `plugins_loaded`).

So we also need to check that the action `wpml_tm_load` is still
attached to `wpml_loaded`.

Also removed unused global in the method.

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-2973