Force to remove `woocommerce` domain from `$l10n_unloaded` global
There's a complex sequence of actions (some before ST is fully
operating) that is preventing the locale to be switched (the MO objects
are not correctly replaced).

When WC is loading its text domain, it also unload it first and we will
be stuck with this during the whole request when the request starts with
a non-english default language and temporarily switch to english.

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-3112