Prevent WPML notices to be wrapped inside the new WC Admin bar
We decided to continue showing all WPML notices for now because we
cannot adjust the CC test properly before the release of WC 4.0.

Later we can think of only showing errors/warnings, and let info notices
to be wrapped.

Also started including Webpack configuration mostly taken from ST. I
tried to install only a part of the npm dependencies but I had troubles
when building the script, so decided to keep most of it (even React
stuff). But anyway we will probably need it some day and it's only for
development purpose.

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-3121