Translate comments with custom MO files instead of DB calls
I was not able to find the place where the strings are registered and maybe this code is not used anymore. This should be investigated in the future.
It might be needed for backward compatibility or for compatibility with
another addon... Anyway, this should fix the performance problem at
least.

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-3245