Fix WC Bookings reminder email language
1. We are now hooking on the `wc-booking-reminder` cron task to set the
language just before to send the email. Note that the
`translate_notification` expects an order ID and it receives a booking
ID we also have a filter to get the language from the order holding the
booking.

2. The second `wc-booking-reminder` callback was obviously not the
correct one, but I think it's not used because it's for admin requests
and it should not be hit in cron tasks.

3. I left some comments to explain some magic and verify some
filters/actions which seems useless now.

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-3268