Fix WC Bookings email strings not showing updated value in WooCommerce -> Settings -> Emails
I tried to follow the same approach that we have for standard emails in
`\WCML_Emails::filter_emails_strings`. For this, I slightly refactored
the code to get the language inside `get_email_translated_string`. We'll
now pass a bool `$isAdminEmail` instead of the language. This method is
already covered with many tests.

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-3279