WIP
https://onthegosystems.myjetbrains.com/youtrack/issue/wpmlga-771
I have copied the stubs from Andrea's MRs (on master) because I have
troubles with stubs generation. The generation is actually OK, but the
analysis is showing fatal errors because some classes are not declared.
It could be a matter of order when we create the stubs and also the
branches of other plugins...