Prevent undue save warning popup in multi-currency settings
We have a standard form element that also contain a partial React
application.

Some inputs of the React application are included in the
`jQuery.serialize` we use to check changes in the form.

I could not find a clean solution to exclude React elements from
serialize, but it should be enough for now.

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-3385