Fix failing integration tests in `Test_WCML_Product_Prices`
I am not sure exactly how these tests could be passing in the past (some
assertions were contradictory with the settings).

I also had to fix the case when we simulate switching the language, I
don't think this is a real case but more a way to test currency
resolution with different "by languages" default currencies.

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-3519