Fix order endpoint showing 404 page with PHP 8
The way we were computing `$uri_vars_not_in_query_vars` before was not
correct and I suspect it was always returning `false`. We were looking
for the array value instead of the array key.

And with a PHP (<8) odd behavior, the `in_array` was always returning
returning `true` (see YT for more details).

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-3553