Remove `pre_get_posts` action to maybe display a 404 on the account page
The action callback is very confusing and after many tests, I could not
find a situation where displaying a 404 was required.

The original ticket were this code was added (5 years ago) is not clear
either.

We introduced a regression in the previous version because on of the
condition was always `false` (and we wrongly fixed it). I think this
code is probably useless now.

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-3574