Refactor Bookings emails
We need to call `\WCML_Emails::wcml_get_translated_email_string` with a
defined user language. We will deprecate this method
because it's confusing and prone to bugs (for example, a booking ID
might be different than the order ID) and we'll use a new method with
relevant parameters (no order ID).

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-3428