Create the WC pages immediately in the correct language
The `wp_insert_post` is not taking into account the current language,
but the global language for the request. And we used to fix it in our
code in a complementary step.

With translate everything mode, jobs are immediately created on
`wpml_after_save_post` if the post is in the default language.

That's why need our created post immediately in the secondary language.

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-3648