Prevent PHP notice when adding a currency switcher widget with WP 5.8 blocks or Elementor Page Builder
First we'll try to get the sidebar `id` from the `args` (legacy), if
missing we'll fallback to the `$instance`.

Eventuallyt, if no sidebar `id` is provided, the display will fallback to the default currency widget.

I also did some reformatting/improvement in the unit test and removed the
`runInSeparateProcess` which was not needed.

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-3510