Fix default currency languages for missing languages
Inside `\WCML_Multi_Currency::init_currencies`, we have some logic to
populate currency languages when the language is missing (i.e. a new
language was added to the site).

Badly, this was not saved to the WCML settings becausewe are changing
the `currency` property only, not the full WCML settings.

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-3691