Trigger `wcml_update_extra_fields` when synchronizing product data
This action hook has been added in the CTE saving callback and it was
missing for ATE.

Unfortunately, we have a very specific code to handle CTE and the hook
relies on submitted data from CTE.

I tried to re-build the expected data inside the product synchronization
logic.

AFAICS, this will fix product bundles and also other 3rd party addons.

Note: I had to find a location to put this hook inside the sync product
logic, but unfortunately the order of the logic is slightly different
with what we have in the CTE callback.

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-3275