Fix product search form block  widget not working in non-default language
This block form is stored in the DB with the action URL in the
default language (differently from the general search block which is
rendered dynamically).

We also need to add the language hidden field when the the site is using
"language as a parameter". But this is handled by
`SitePress::get_search_form_filter`.

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-3720