Add full ATE support for WC Bookings
Some notes:
1. Once the translation is completed (applied), we need to synchronize
the product. This was done separately for CTE in
`\WCML_Editor_UI_Product_Job::save_translations`.
2. In `\WCML_Bookings::save_booking_data_to_translation`, we'll check
the booking type against the original because the type is not set yet
for the translation at this point. The private methods
`save_person_translation` and `save_resource_translation` are
exclusively called from there (that's why I moved the condition).

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-3789