Do not load `WCML\Multicurrency\UI\Hooks` if dependency check is `false`
The problem happens on a site with WC, ST and WCML, when accessing the
main WCML menu. It's an edge case for a temporary state.

Our `isStandAlone` function relies on Core only, but I prefer not to
change it now, we might introduce other side effects.

In this commit, I also made `\WCML_Dependencies` a shared instance so we
stop getting it from `woocommerce_wpml` instance.

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-3972