Provide more information when the exchange service is returning an error
The way we format the error message will be uniform for all services.
Basically we try to gather all possible codes and messages and display
it in a concatenated string.

The reason is that each service has a different error signature and it
does not even respect their own documentation. For one service, the
signature might change depending on the error (using different keys or
paths).

At least we are providing as much information as we can to the user.

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-3996