Sort exchange rate services by alphabetical order and use the first as default
The first one was "fixer.io" which was immediately showing a warning not
very appealing... In order to not have an arbitrary order, I decided to
opt for an alphabetical order which also puts "currencylayer" on top.
And from my recent tests, I used currency layer without any problem.

We'll improve this order and more generally our strategy in presenting
exchange rate service in a next version.

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-3975#focus=Comments-102-510622.0-0