Prevent fatal error with 3rd party gateways classes not loaded
In the reported cases, the gateway class is not autoloaded and it will
be loaded manually just after enabling it which is after our initialization
process. On the next request, the class definition will be loaded
properly.

We should just skip the case when a class is not defined to make our
code more resilient.

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-4034