Fix showing average rating on single product page.
This is a defect of my previous commit. We always show the average rating which is the average in all languages. But this is only displayed if rating_count is greater than zero. For this reason, we want to always filter this value, summing up ratings on both languages.

Not to be confused with reviews_count, which has different count per language.

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-4147/Rating-stars-are-not-synced-between-languages