[wcml-4133] Show empty categories when it has products in the source language with `display as translated` mode
It's a partial fix.

As the category count needs to greater than 0 for the category to
display, we will get the count from the source in that case.
If the source count is greater than 0, we'll take it for the translation
and it will display.
For translated categories where the count is already greater than 0, the
count might be incorrect, but the category will show anyway (which is
the most important).

Note that this fix works only when the filter hook `woocommerce_product_subcategories_args` is fired (shop home with categories).
We can extend that to other case in the future.

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-4133/