[wcml-4367] Temporarily ignore PHPStan error "Cannot assign offset 'lang' to WP_REST_Request"
This error poped up as soon as I updated Installer because the dev
dependencies in `vendor-bin` where also updated (as a side effect).

`WP_REST_Request` implements `ArrayAccess` and we are not assigning in
our code. It would be a bug in one of the updated dependencies (PHPStan,
Stubs, ...).

For now I will just leave this error in the baseline.

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-4367/Update-Installer-version-to-3.1.0