Attempt attribute translation only if they exist
We recently started processing attributes with an empty value as if it was an attribute with ANY selected.
The problem is that some meta keys are not attributes, and get_post_meta doesn't distinguish between a non-existant meta or an empty one.
Using metadata_exists() we can check if it actually exists before processing it.

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-4472/HPOS-Switching-language-in-order-screen-produces-error