[wcml-4491] HPOS: Use CRUD for order currency (1st iteration)
In this 1st iteration, I've just replaced `get_post_meta` and
`update_post_meta` with the new methods
`\WCML\Orders\Helper::getCurrency` and
`\WCML\Orders\Helper::setCurrency`.

THough, using the CRUD approach requires some refined logic because in
some places we have cross-dependencies between the instantion of the
order object and `\WCML\Orders\Helper::getCurrency`. Moreover, we cannot call `wc_get_order` too early.
For these 2 reasons, I had to use extra logic with direct DB query.

In the next iteration, we'll have to take care of DB queries directly
filtered on `posts` and `postmeta` tables.

Note: I guess there's a bug in `WCML_Orders::order_language_dropdown`
because it was checking on the order currency instead of the order language.
As I don't really know the possible effect of fixing this bug, I don't
want to fix it for now, but I've added a comment in the code.

https://onthegosystems.myjetbrains.com/youtrack/issue/wcml-4491/